requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
duckdb>=0.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Pipeline and orchestration
dagster>=1.6.0
dagster-webserver>=1.6.0
dagster-graphql>=1.6.0

# LLM and embeddings
langchain-core>=0.1.27
langchain-openai>=0.0.8
langchain-community>=0.0.24

# API and services
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
click>=8.1.7

# Caching and persistence
diskcache>=5.6.3
joblib>=1.3.2
cachetools>=5.3.2

# Pattern matching and parsing
pygrok>=1.0.0
regex>=2023.12.25

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
dynaconf>=3.2.4

# Development dependencies
pytest>=7.4.4
pytest-cov>=4.1.0
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
pylint>=3.0.3

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
mkdocstrings>=0.24.0 